The detail.view and Auction.view files have been modified.
The buy button found in detail.vew, disappears in case the token is not available, and if the tokenId is related to an auction, it redirects to the page where the auction is located.

In aution.view the problem that was had when reloading the page was solved. Added a hook that stores the highest bidder.
    and the lowest bid. If the tokenId is related to a sale or if it is not available, redirect to the details page